+++ R/pairwise.r
@@ -19,6 +19,17 @@
 #' slope vectors, this is the difference in location between estimated change 
 #' for the dependent variables, per one-unit change of the covariate considered.
 #' For univariate slopes, this is the absolute difference between slopes.}
+#' \item{\bold{Standardized distance between vectors, "stdist"}}{Same as the 
+#' distance between vectors, but distances are divided by the model standard error 
+#' (square-root of the trace of the residual covariance matrix, adjusted by sample size). 
+#' Pairwise tests with this statistic should be consistent with ANOVA results.}
+#' \item{\bold{Mahalanobis distance between vectors, "mdist"}}{Similar to the
+#' standardized distance between vectors but the inverse of the 
+#' residual covariance matrix is used in calculation of the distance, rather than
+#' dividing the Euclidean distance between means and dividing by the standard error. 
+#' If the residual covariance matrix is singular, Mahalanobis distances will not be 
+#' calculated.  Pairwise tests with this statistic should be consistent with 
+#' MANOVA results.}
 #' \item{\bold{Vector correlation, "VC"}}{ If LS mean or slope vectors are 
 #' scaled to unit size, the vector correlation is the inner-product of the 
 #' scaled vectors.
@@ -48,7 +59,8 @@
 #' 
 #' The \code{\link{summary.pairwise}} function is used to select a test 
 #' statistic for the statistics described above, as
-#' "dist", "VC", "DL", and "var", respectively.  If vector correlation is tested, the \code{angle.type} argument can be used to choose between radians and
+#' "dist", "VC", "DL", and "var", respectively.  If vector correlation is tested, 
+#' the \code{angle.type} argument can be used to choose between radians and
 #' degrees.
 #' 
 #' The null model is defined via \code{\link{lm.rrpp}}, but one can 
@@ -87,6 +99,25 @@
 #'  and subsequent versions.
 #' }
 #' 
+#'  \subsection{Notes for RRPP 2.0.4 and subsequent versions}{ 
+#'  The test types, standardized distance between vectors, "stdist", and Mahalanobis distances
+#'  between vectors were added.  The former
+#'  simply divides the distance between vectors by model standard error (square-root 
+#'  of the trace of the residual covariance matrix, adjusted by sample size).  
+#'  This is a multivariate generalization
+#'  of a t-statistic for the difference between means.  It is not the same as Hotelling 
+#'  squared-T-statistic, which requires incorporating the inverse of the residual covariance 
+#'  matrix in the calculation of the distance, a calculation that also requires a non-singular
+#'  covariance matrix.  However, the Mahalanobis distances are similar (and proportional) 
+#'  to the Hotelling squared-T-statistic. Pairwise tests using Mahalanobis distances represent
+#'  a non-parametric analog to the parametric Hotelling squared-T test.  Both tests should be 
+#'  better for GLS model fits compared
+#'  to Euclidean distances between means, as the total sums of squares are more likely to vary
+#'  across random permutations. In general, if ANOVA is 
+#'  performed a pairwise test with "stdist" is a good association; if MANOVA is performed,
+#'  a pairwise test with "mdist" is a good association.
+#' }
+#'  
 #' @param fit A linear model fit using \code{\link{lm.rrpp}}.
 #' @param fit.null An alternative linear model fit to use as a null 
 #' model for RRPP, if the null model
@@ -108,6 +139,11 @@
 #' would be part of the 
 #' model fit, but it is not necessary for that to be the case in order 
 #' to obtain results.
+#' @param verbose A logical value for whether to include extra results, specifically 
+#' the Mahalanobis distances among means, which requires the calculation of residual 
+#' covariance matrices for each permutation.  This should be generally
+#' FALSE, unless Mahalanobis distances are desired, in which case it must be TRUE.  
+#' Verbose computations can require much additional time.
 #' @param print.progress If a null model fit is provided, a logical 
 #' value to indicate whether analytical 
 #' results progress should be printed on screen.  Unless large data 
@@ -121,6 +157,8 @@
 #' \item{LS.means}{LS means for groups, across permutations.}
 #' \item{slopes}{Slopes for groups, across permutations.}
 #' \item{means.dist}{Pairwise distances between means, across permutations.}
+#' \item{std.means.dist}{Pairwise distances between means, across permutations, standardized.}
+#' \item{mah.means.dist}{Pairwise Mahalanobis distances between means, across permutations.}
 #' \item{means.vec.cor}{Pairwise vector correlations between 
 #' means, across permutations.}
 #' \item{means.lengths}{LS means vector lengths, by group, across permutations.}
@@ -166,7 +204,7 @@
 #' 
 #' Pupfish$Y <- ordinate(Pupfish$coords)$x[, 1:3]
 #' 
-#' ## Pairwise comparisons of LS means
+#' ## Pairwise comparisons among LS means
 #' 
 #' # Note: one should increase RRPP iterations but a 
 #' # smaller number is used here for demonstration 
@@ -183,9 +221,15 @@
 #' PW1 <- pairwise(fit1, groups = pup.group)
 #' PW1
 #' 
-#' # distances between means
+#' # distances among means
 #' summary(PW1, confidence = 0.95, test.type = "dist") 
 #' summary(PW1, confidence = 0.95, test.type = "dist", stat.table = FALSE)
+#' 
+#' # standardized distances among means
+#' summary(PW1, confidence = 0.95, test.type = "stdist") 
+#' 
+#' # Mahalanobis (generalized) distances among means
+#' summary(PW1, confidence = 0.95, test.type = "mdist") 
 #' 
 #' # absolute difference between mean vector lengths
 #' summary(PW1, confidence = 0.95, test.type = "DL") 
@@ -227,11 +271,23 @@
 #' }
 #' 
 pairwise <- function(fit, fit.null = NULL, groups, covariate = NULL, 
-                     print.progress = FALSE) {
+                     verbose = FALSE, print.progress = FALSE) {
   fitf <- fit
   term.labels <- fit$LM$term.labels
   n <- fit$LM$n
   p <- fit$LM$p
+  
+  res.df <- fit$ANOVA$df[(length(fit$ANOVA$df) - 1)]
+  SE <- sqrt(fit$ANOVA$RSS.model[NROW(fit$ANOVA$RSS.model), ] / res.df)
+  
+  if(verbose){
+    if(res.df < fit$LM$p) {
+      cat("\nVerbose results were requested but the residual covariance matrix is singular.\n",
+          "Therefore, residual covariance matrices will not be estimated for every random permutation.\n")
+      verbose = FALSE
+    }
+  }
+  
   PermInfo <- getPermInfo(fit, attribute = "all")
   Models <- getModels(fit, attribute = "all")
   ind <- PermInfo$perm.schedule
@@ -316,7 +372,7 @@ pairwise <- function(fit, fit.null = NULL, groups, covariate = NULL,
   H <- tcrossprod(solve(Qf$R), Qf$Q)
   getCoef <- function(y) H %*% y
   
-  if(is.null(fitf$LM$random.coef)) {
+  if(is.null(fitf$LM$random.coef) || !is.null(fit.null)) {
     
     coef.n <- lapply(1:perms, function(j){
       step <- j
@@ -339,7 +395,7 @@ pairwise <- function(fit, fit.null = NULL, groups, covariate = NULL,
   }
   
   groups <- factor(groups)
-  gp.rep <- by(groups, groups, length)
+  n_groups <- tapply(groups, groups, length)
   
   
   if(length(groups) != n) 
@@ -361,10 +417,63 @@ pairwise <- function(fit, fit.null = NULL, groups, covariate = NULL,
     means <- getLSmeans(fit = fitf, g = groups)
     slopes <- NULL
     means.dist <- lapply(means, function(x) as.matrix(dist(x)))
+    pw_matches <- combn(n_groups, 2, simplify = FALSE)
+    pw.sample.adjust <- lapply(pw_matches, function(x) 1/x[1] + 1/x[2])
+    ng <- length(n_groups)
+    pw_correct <- as.dist(matrix(0, ng, ng))
+    pw_correct[1:length(pw.sample.adjust)] <- sqrt(unlist(pw.sample.adjust))
+    pw_correct <- as.matrix(pw_correct)
+    diag(pw_correct) <- 1
+    std.means.dist <- Map(function(m, s) m/(s * pw_correct), means.dist, as.list(SE))
     means.vec.cor <- lapply(means, vec.cor.matrix)
     means.length <- lapply(means, function(x) sqrt(rowSums(x^2)))
     means.diff.length <- lapply(means.length, function(x) 
       as.matrix(dist(as.matrix(x))))
+    if(verbose){
+      if(print.progress){
+        cat(paste("\nEstimation of residual covariance matrices:", 
+                  perms, "permutations.\n"))
+        pb <- txtProgressBar(min = 0, max = perms+1, initial = 0, style=3)
+      }
+      
+      ResCovs <- lapply(1:perms, function(j){
+        rrpp.args$ind.i <- ind[[j]]
+        y <- do.call(rrpp, rrpp.args)
+        yh <- fastFit(Qf$Q, y, n, p)
+        R <- y - yh
+        step <- j
+        if(print.progress) 
+          setTxtProgressBar(pb,step)
+        crossprod(R) / res.df
+      })
+      
+      step <- perms + 1
+      if(print.progress) {
+        setTxtProgressBar(pb,step)
+        close(pb)
+      }
+        
+      if(print.progress){
+        cat(paste("\nCalculation of pairwise Mahalanobis distances:", 
+                  perms, "permutations.\n"))
+        pb <- txtProgressBar(min = 0, max = perms+1, initial = 0, style=3)
+      }
+      
+      mah.means.dist <- lapply(1:perms, function(j){
+        m <- means[[j]]
+        s <- ResCovs[[j]]
+        step <- j
+        md <- mahal_dist(m, s, method = "euclidean")
+        if(print.progress) 
+          setTxtProgressBar(pb,step)
+        as.matrix(md)
+      })
+      step <- perms + 1
+      if(print.progress) {
+        setTxtProgressBar(pb,step)
+        close(pb)
+      }
+    } else mah.means.dist <- NULL
     
     slopes.length <- NULL
     slopes.dist <- NULL
@@ -376,6 +485,8 @@ pairwise <- function(fit, fit.null = NULL, groups, covariate = NULL,
     means <- NULL
     slopes <- getSlopes(fit = fitf, x = covariate, g = groups)
     means.dist <- NULL
+    std.means.dist <- NULL
+    mah.means.dist <- NULL
     means.vec.cor <- NULL
     means.length <- NULL
     means.diff.length <- NULL
@@ -405,13 +516,15 @@ pairwise <- function(fit, fit.null = NULL, groups, covariate = NULL,
   colnames(vars) <- c("obs", paste("iter", 1:(perms -1), sep = "."))
   
   out <- list(LS.means = means, slopes = slopes, 
-              means.dist = means.dist, means.vec.cor = means.vec.cor, 
+              means.dist = means.dist, std.means.dist = std.means.dist,
+              mah.means.dist = mah.means.dist, 
+              means.vec.cor = means.vec.cor, 
               means.length = means.length, 
               means.diff.length = means.diff.length,
               slopes.dist = slopes.dist, slopes.vec.cor = slopes.vec.cor,
               slopes.length = slopes.length, 
               slopes.diff.length = slopes.diff.length,
-              vars = vars, n = n, p = p, PermInfo = fitf$PermInfo)
+              vars = vars, SE = SE, n = n, p = p, PermInfo = fitf$PermInfo)
   
   class(out) <- "pairwise"
   out